move record utils to record pkg
