+++ integration_tests/latest_jdk_tests/src/test/java/io/fury/integration_tests/RecordSerializersTest.java
@@ -23,8 +23,8 @@ import io.fury.Fury;
 import io.fury.resolver.MetaContext;
 import io.fury.serializer.CompatibleMode;
 import io.fury.test.bean.Struct;
-import io.fury.util.RecordComponent;
-import io.fury.util.RecordUtils;
+import io.fury.util.record.RecordComponent;
+import io.fury.util.record.RecordUtils;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
