+++ java/fury-core/src/main/java/io/fury/builder/CodecBuilder.java
@@ -43,9 +43,9 @@ import io.fury.resolver.ClassInfo;
 import io.fury.resolver.ClassInfoCache;
 import io.fury.type.Descriptor;
 import io.fury.util.Platform;
-import io.fury.util.RecordUtils;
 import io.fury.util.ReflectionUtils;
 import io.fury.util.StringUtils;
+import io.fury.util.record.RecordUtils;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 import java.util.HashMap;
