+++ java/fury-core/src/main/java/io/fury/serializer/CompatibleSerializer.java
@@ -25,9 +25,9 @@ import io.fury.resolver.FieldResolver;
 import io.fury.resolver.RefResolver;
 import io.fury.util.FieldAccessor;
 import io.fury.util.Platform;
-import io.fury.util.RecordInfo;
-import io.fury.util.RecordUtils;
 import io.fury.util.ReflectionUtils;
+import io.fury.util.record.RecordInfo;
+import io.fury.util.record.RecordUtils;
 import java.lang.invoke.MethodHandle;
 import java.util.Arrays;
 import java.util.Collection;
