+++ java/fury-core/src/main/java/io/fury/serializer/MetaSharedSerializer.java
@@ -33,9 +33,9 @@ import io.fury.type.DescriptorGrouper;
 import io.fury.type.Generics;
 import io.fury.util.FieldAccessor;
 import io.fury.util.Platform;
-import io.fury.util.RecordInfo;
-import io.fury.util.RecordUtils;
 import io.fury.util.ReflectionUtils;
+import io.fury.util.record.RecordInfo;
+import io.fury.util.record.RecordUtils;
 import java.lang.invoke.MethodHandle;
 import java.lang.reflect.Field;
 import java.util.ArrayList;
