+++ java/fury-core/src/main/java/io/fury/serializer/ObjectSerializer.java
@@ -35,9 +35,9 @@ import io.fury.type.GenericType;
 import io.fury.type.Generics;
 import io.fury.util.FieldAccessor;
 import io.fury.util.Platform;
-import io.fury.util.RecordInfo;
-import io.fury.util.RecordUtils;
 import io.fury.util.ReflectionUtils;
+import io.fury.util.record.RecordInfo;
+import io.fury.util.record.RecordUtils;
 import java.lang.invoke.MethodHandle;
 import java.util.ArrayList;
 import java.util.Arrays;
