+++ java/fury-core/src/main/java/io/fury/type/Descriptor.java
@@ -27,9 +27,9 @@ import io.fury.annotation.Ignore;
 import io.fury.annotation.Internal;
 import io.fury.collection.Tuple2;
 import io.fury.util.Platform;
-import io.fury.util.RecordComponent;
-import io.fury.util.RecordUtils;
 import io.fury.util.StringUtils;
+import io.fury.util.record.RecordComponent;
+import io.fury.util.record.RecordUtils;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
