+++ java/fury-core/src/main/java/io/fury/util/record/RecordComponent.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.fury.util;
+package io.fury.util.record;
 
 import java.lang.reflect.Method;
 import java.lang.reflect.Type;
